Cambios generales estructurales

+++ Dtos/Articulo/StockArticuloDto.cs
@@ -3,8 +3,8 @@
     public class StockArticuloDto
     {
         public  long nStock { get; set; }
-        public int stockSeguridad { get; set; }
-        public int stockActual { get; set; }
+        public long stockSeguridad { get; set; }
+        public long stockActual { get; set; }
         public DateTime fechaStockInicio { get; set; }
         public DateTime fechaStockFin { get; set; }
         public long idArticulo { get; set; }
